Added a script for "Structrue Search"
This essentially loads the basic types from a code list, then uses the offsets within that list to build a pattern to search for. If you know three or four of the values you are looking for, you can usually narrow it down to the exact addresses you want with a single search.